+++ PMSlib/app/src/main/java/br/ufc/mdcc/cmu/pmslib/exception/BrokerMQTTException.java
@@ -1,0 +1,7 @@
+package br.ufc.mdcc.cmu.pmslib.exception;
+
+public class BrokerMQTTException extends Exception {
+    public BrokerMQTTException(String message) {
+        super(message);
+    }
+}
